Prep v228: Add remaining updates from upstream (2/3)
Apply remaining fixes and the performed move of utility functions
into their own foo-util.[hc] files on libbasic.
