+++ src/basic/bus-label.c
@@ -21,10 +21,11 @@
 
 #include <stdlib.h>
 
-#include "util.h"
-#include "macro.h"
-
+#include "alloc-util.h"
 #include "bus-label.h"
+#include "hexdecoct.h"
+#include "macro.h"
+#include "util.h"
 
 char *bus_label_escape(const char *s) {
         char *r, *t;
