+++ src/basic/fileio-label.h
@@ -27,5 +27,3 @@
 
 int write_string_file_atomic_label(const char *fn, const char *line);
 // UNNEEDED int write_env_file_label(const char *fname, char **l);
-// UNNEEDED int fopen_temporary_label(const char *target,
-//                          const char *path, FILE **f, char **temp_path);
