+++ src/basic/login-util.c
@@ -19,8 +19,9 @@
   along with systemd; If not, see <http://www.gnu.org/licenses/>.
 ***/
 
-#include "login-util.h"
 #include "def.h"
+#include "login-util.h"
+#include "string-util.h"
 
 bool session_id_valid(const char *id) {
