+++ src/basic/memfd-util.h
@@ -21,7 +21,8 @@
   along with systemd; If not, see <http://www.gnu.org/licenses/>.
 ***/
 
-
+#include <sys/types.h>
+#include <inttypes.h>
 
 int memfd_new(const char *name);
 // UNNEEDED int memfd_new_and_map(const char *name, size_t sz, void **p);
