+++ src/basic/utf8.c
@@ -44,11 +44,13 @@
  */
 
 #include <errno.h>
-#include <stdlib.h>
 #include <inttypes.h>
-#include <string.h>
 #include <stdbool.h>
+#include <stdlib.h>
+#include <string.h>
 
+#include "alloc-util.h"
+#include "hexdecoct.h"
 #include "utf8.h"
 #include "util.h"
