+++ src/basic/virt.h
@@ -48,6 +48,7 @@ enum {
         VIRTUALIZATION_LXC,
         VIRTUALIZATION_OPENVZ,
         VIRTUALIZATION_DOCKER,
+        VIRTUALIZATION_RKT,
         VIRTUALIZATION_CONTAINER_OTHER,
         VIRTUALIZATION_CONTAINER_LAST = VIRTUALIZATION_CONTAINER_OTHER,
 
@@ -66,6 +67,8 @@ static inline bool VIRTUALIZATION_IS_CONTAINER(int x) {
 int detect_vm(void);
 int detect_container(void);
 // UNNEEDED int detect_virtualization(void);
+
+int running_in_chroot(void);
 
 const char *virtualization_to_string(int v) _const_;
 int virtualization_from_string(const char *s) _pure_;
